requirements.txt
python-telegram-bot[job-queue]
python-dotenv==1.1.1
